movies list completed, movie details screen partially completed

+++ app/src/main/java/in/movies/niranjan/com/movieproj/models/MovieDetail.java
@@ -44,7 +44,7 @@ public class MovieDetail {
     public String overview;
 
     @SerializedName("popularity")
-    public Double popularity;
+    public float popularity;
 
     @SerializedName("poster_path")
     public String posterPath;
